split user class from auth class
